--- src/components/pages/Blog/BlogItem/index.ts
@@ -1,1 +1,0 @@
-export * from './BlogItem'
