+++ src/components/pages/Blog/BlogList/generateTitle.tsx
